Backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# Research & Data Tools (for Jan Aushadhi)
fuzzywuzzy==0.18.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# AI & LLM Integration
langchain-google-genai==3.2.0
langchain-core==1.1.0

# Utilities
requests==2.32.3
pydantic==2.9.2
httpx==0.27.0